Handle Y endstop noise on homing
Home sometimes failed, and it appears to be due to noise on the Y
endstop. This makes sense, as the endstop lines run right over the
Y axis motor, and couls cause noise. I was finding that homing would
fail after the Y axis was pulsed during a homing operation.